feat: Cleaning up old code
